Code cleanup / refactor
