+++ packages/front-end/services/rule-conflicts.tsx
@@ -1,7 +1,12 @@
+import { useState, type ReactElement } from "react";
+import { Flex } from "@radix-ui/themes";
 import { FeatureRule } from "shared/types/feature";
 import { ExperimentInterfaceStringDates } from "shared/types/experiment";
 import { extractConditionAttributeKeys } from "shared/util";
 import { paddedVersionString } from "@growthbook/growthbook";
+import Callout from "@/ui/Callout";
+import Link from "@/ui/Link";
+import Text from "@/ui/Text";
 import { getUpcomingScheduleRule } from "@/services/scheduleRules";
 import { isRuleInactive } from "@/services/features";
 
@@ -893,4 +898,110 @@ export function getRuleReachability(
   }
 
   return result;
+}
+
+// ---------------------------------------------------------------------------
+// Conflict display (UI)
+// ---------------------------------------------------------------------------
+
+// Targeting conflicts resolved for display (rule ids → visible rule numbers).
+export type RuleConflictInfo = {
+  // Precise: a rule above fully serves these targeted values ("will not reach").
+  hard: { ruleNumber?: number; attr: string | null; label: string | null }[];
+  // Soft: a rule above also targets this attribute ("may not reach").
+  soft: { attr: string | null; ruleNumbers: number[] }[];
+};
+
+function hardTargetingPhrase(attr: string, label: string): string {
+  const trimmed = label.trim();
+  if (/^[>≥<≤]/.test(trimmed) || /, [<≥]/.test(label)) {
+    return `${attr} ${label}`;
+  }
+  if (label.includes(", ")) {
+    return `${attr} is one of ${label}`;
+  }
+  return `${attr} is ${label}`;
+}
+
+function hardSentence(c: {
+  ruleNumber?: number;
+  attr: string | null;
+  label: string | null;
+}): string {
+  const ref = c.ruleNumber ? `Rule ${c.ruleNumber}` : "An earlier rule";
+  if (c.label === null) {
+    return c.attr
+      ? `${ref} already serves all traffic matching this rule's ${c.attr} targeting before it reaches this rule.`
+      : `${ref} already serves all traffic before it reaches this rule.`;
+  }
+  if (c.attr) {
+    return `${ref} already serves traffic where ${hardTargetingPhrase(c.attr, c.label)} before it reaches this rule.`;
+  }
+  return `${ref} already serves traffic matching "${c.label}" before it reaches this rule.`;
+}
+
+function softSentence(c: {
+  attr: string | null;
+  ruleNumbers: number[];
+}): string {
+  const nums = c.ruleNumbers.filter((n) => n > 0);
+  const refs = nums.length
+    ? nums.map((n) => `Rule ${n}`).join(", ")
+    : "An earlier rule";
+  // `attr` null → an untargeted partial rollout above siphons traffic.
+  if (c.attr === null) {
+    const verb = nums.length > 1 ? "serve" : "serves";
+    return `${refs} ${verb} a share of all traffic before it reaches this rule, so some matching traffic may not reach it.`;
+  }
+  const verb = nums.length > 1 ? "target" : "targets";
+  return `${refs} also ${verb} ${c.attr}, so some matching traffic may be served there first.`;
+}
+
+// Generic warning that some/all targeted traffic won't (or may not) reach this
+// rule, with an expandable explanation of which rule(s) consume it.
+export function ConflictCallout({
+  isUnreachable,
+  conflicts,
+}: {
+  isUnreachable: boolean;
+  conflicts: RuleConflictInfo;
+}): ReactElement {
+  const [open, setOpen] = useState(false);
+  const hasHard = conflicts.hard.length > 0;
+  const hasSoft = conflicts.soft.length > 0;
+  const headline = isUnreachable
+    ? "No matching traffic will reach this rule."
+    : hasHard
+      ? "Some matching traffic will not reach this rule."
+      : "Some matching traffic may not reach this rule.";
+  const hasDetails = hasHard || hasSoft;
+  const status = isUnreachable ? "error" : "warning";
+  return (
+    <Callout status={status} size="sm" contentsAs="div">
+      <Flex direction="column" gap="1">
+        <Flex align="center" gap="2" wrap="wrap">
+          <span>{headline}</span>
+          {hasDetails && (
+            <Link role="button" onClick={() => setOpen((o) => !o)}>
+              {open ? "Hide details" : "See details"}
+            </Link>
+          )}
+        </Flex>
+        {open && hasDetails && (
+          <Flex mt="1" direction="column" gap="1">
+            {conflicts.hard.map((c, i) => (
+              <Text as="div" size="small" key={`h${i}`}>
+                - {hardSentence(c)}
+              </Text>
+            ))}
+            {conflicts.soft.map((c, i) => (
+              <Text as="div" size="small" key={`s${i}`}>
+                - {softSentence(c)}
+              </Text>
+            ))}
+          </Flex>
+        )}
+      </Flex>
+    </Callout>
+  );
 }
